@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "seed-salesforce"
-version = "0.1.0"
+version = "0.1.1"
 description = "Package for connecting SEED data to Salesforce"
 authors = ["Nicholas Long <nicholas.long@nrel.gov>", "Katherine Fleming <katherine.fleming@nrel.gov>"]
 license = "BSD-3-Clause"
@@ -14,27 +14,26 @@ keywords = ["SEED Platform", "Building Benchmarking", "Salesforce", "Building Pe
 classifiers = [
     "Development Status :: 4 - Beta",
     "Intended Audience :: Developers",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
 ]
 packages = [
-    { include = "seed_salesforce"}
+    { include = "seed_salesforce" }
 ]
 
 [tool.poetry.dependencies]
-python = ">=3.8.1, <3.11"
-python-dateutil = "> 0"
-
-simple-salesforce = "^1.11.6"
+python = ">=3.9, <3.13"
+python-dateutil = "*"
+simple-salesforce = "^1.12.6"
 
 [tool.poetry.dev-dependencies]
-autopep8 = "2.0.1"
-flake8 = "6.0.0"
-mypy = "^1.0.1"
-pre-commit = "3.0.4"
+mypy = "^1.11.2"
+pre-commit = "^3.8.0"
 pytest = "^7.1.2"
 pytest-cov = "^4.0.0"
+ruff = "^0.6.7"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
