@@ -2,11 +2,11 @@
 
 # -*- coding: utf-8 -*-
 """
-    Dummy conftest.py for alfalfa.
+Dummy conftest.py for seed-salesforce.
 
-    If you don't know what this is for, just leave it empty.
-    Read more about conftest.py under:
-    https://pytest.org/latest/plugins.html
+If you don't know what this is for, just leave it empty.
+Read more about conftest.py under:
+https://pytest.org/latest/plugins.html
 """
 
 # import pytest
