@@ -4,30 +4,28 @@ import random
 import unittest
 from pathlib import Path
 
-import pytest
-
 from seed_salesforce.salesforce_client import SalesforceClient
 
 
 class SalesforceClientTest(unittest.TestCase):
     def test_init_by_file(self):
-        config_file = Path('salesforce-config-dev.json')
+        config_file = Path("salesforce-config-dev.json")
         sf = SalesforceClient(connection_config_filepath=config_file)
         # check if the Salesforce connection has the base_url method and that it is not None
-        assert hasattr(sf.connection, 'base_url')
+        assert hasattr(sf.connection, "base_url")
         # print(f"sf connection: {sf.connection.base_url}")
         assert sf.connection.base_url is not None
 
 
 class SalesforceIntegrationTest(unittest.TestCase):
     def setUp(self) -> None:
-        self.output_dir = Path('tests/output')
+        self.output_dir = Path("tests/output")
         if not self.output_dir.exists():
             self.output_dir.mkdir()
 
         # create the connection to salesforce.
         # These tests run against a development version of Salesforce
-        config_file = Path('salesforce-config-dev.json')
+        config_file = Path("salesforce-config-dev.json")
         self.sf = SalesforceClient(connection_config_filepath=config_file)
 
         self.test_org = "New Test Account"
@@ -41,29 +39,29 @@ class SalesforceIntegrationTest(unittest.TestCase):
     # @pytest.mark.skip(reason="Need to confirm that this is a valid test")
     def test_search_and_create_and_delete_account(self):
         r = random.randrange(1, 5000, 2)
-        test_account_name = 'Scrumptious Ice Cream' + str(r)
+        test_account_name = f"Scrumptious Ice Cream{r}"
         account = self.sf.find_account_by_name(test_account_name)
         # The first time should not have any results
         assert not account
 
         details = {
-            'Type': 'Ice Cream Shop',
-            'Phone': '555-867-5309',
-            'Website': 'http://www.scrumptiousco.com',
+            "Type": "Ice Cream Shop",
+            "Phone": "555-867-5309",
+            "Website": "http://www.scrumptiousco.com",
         }
         if self.account_record_type:
-            details['RecordTypeID'] = self.account_record_type
+            details["RecordTypeID"] = self.account_record_type
 
         account = self.sf.create_account(test_account_name, **details)
-        account_id = account['Id']
+        account_id = account["Id"]
         # print(f"ACCOUNT ID: {account_id}")
-        assert account['Name'] == test_account_name
-        assert account['Type'] == 'Ice Cream Shop'
+        assert account["Name"] == test_account_name
+        assert account["Type"] == "Ice Cream Shop"
         assert account_id is not None
 
         # Find the record
         account = self.sf.find_account_by_name(test_account_name)
-        assert account['Id'] == account_id
+        assert account["Id"] == account_id
 
         # now delete it to cleanup
         success = self.sf.delete_account_by_id(account_id)
@@ -72,46 +70,49 @@ class SalesforceIntegrationTest(unittest.TestCase):
     # @pytest.mark.skip(reason="Need to confirm that this is a valid test")
     def test_update_contact_and_account(self):
         r = random.randrange(1, 5000, 2)
-        test_account_name = 'A Fake Account' + str(r)
+        test_account_name = "A Fake Account" + str(r)
         details = {
-            'Type': 'Office',
-            'Phone': '555-867-5309',
-            'Website': 'http://www.scrumptiousco.com'
+            "Type": "Office",
+            "Phone": "555-867-5309",
+            "Website": "http://www.scrumptiousco.com",
         }
         if self.account_record_type:
-            details['RecordTypeID'] = self.account_record_type
+            details["RecordTypeID"] = self.account_record_type
 
         account = self.sf.create_account(test_account_name, **details)
-        account_id = account['Id']
+        account_id = account["Id"]
 
-        success = self.sf.update_account_by_id(account_id, {
-            'Phone': '444-444-4444',
-            'Website': 'http://www.UpdatedWebsite.com'
-        })
+        success = self.sf.update_account_by_id(
+            account_id,
+            {
+                "Phone": "444-444-4444",
+                "Website": "http://www.UpdatedWebsite.com",
+            },
+        )
         # this returns a 204 but seems to do so regardless of whether property exists
         assert success
 
         # contacts
         # create contact (associated with account)
         details = {
-            'AccountId': account_id,
-            'LastName': 'Richard Hendrick' + str(r)
+            "AccountId": account_id,
+            "LastName": "Richard Hendrick" + str(r),
         }
         if self.contact_record_type:
-            details['RecordTypeID'] = self.contact_record_type
+            details["RecordTypeID"] = self.contact_record_type
 
-        self.sf.create_contact('user' + str(r) + '@company.com', **details)
+        self.sf.create_contact("user" + str(r) + "@company.com", **details)
         # retrieve contact
-        contact = self.sf.find_contact_by_email('user' + str(r) + '@company.com')
-        assert contact['AccountId'] == account_id
+        contact = self.sf.find_contact_by_email("user" + str(r) + "@company.com")
+        assert contact["AccountId"] == account_id
 
         # update contact (can't change email)
         details = {
-            'AccountId': account_id,
-            'LastName': 'Russ Hanneman' + str(r)
+            "AccountId": account_id,
+            "LastName": "Russ Hanneman" + str(r),
         }
-        self.sf.create_or_update_contact_on_account('user' + str(r) + '@company.com', **details)
-        assert contact['AccountId'] == account_id
+        self.sf.create_or_update_contact_on_account("user" + str(r) + "@company.com", **details)
+        assert contact["AccountId"] == account_id
 
         # now delete it to cleanup
         success = self.sf.delete_account_by_id(account_id)
@@ -159,34 +160,33 @@ class SalesforceIntegrationTest(unittest.TestCase):
     #     prop_updated2 = self.sf.update_property(property_id, **args)
 
     def test_benchmark(self):
-
         # get first benchmark (to get an Id)
         print(" ...get first benchmark (to get an ID)...")
         benchmark = self.sf.get_first_benchmark()
         print(f" Benchmark: {benchmark}")
         assert benchmark is not None
-        benchmark_id = benchmark['Id']
-        salesforce_benchmark_id = benchmark['Salesforce_Benchmark_ID__c']
+        benchmark_id = benchmark["Id"]
+        salesforce_benchmark_id = benchmark["Salesforce_Benchmark_ID__c"]
 
         # can you retrieve by "Salesforce Benchmark ID" custom field?
         print(" ...retrieving benchmark by Salesforce Benchmark ID...")
         bench_by_custom_id = self.sf.get_benchmark_by_custom_id(salesforce_benchmark_id)
         print(f" benchmark by custom id: {bench_by_custom_id}")
-        assert bench_by_custom_id['Id'] == benchmark_id
+        assert bench_by_custom_id["Id"] == benchmark_id
 
         # can you update a benchmark field?
         print(" ...updating benchmark...")
         # TODO: again assuming this field exists...it might not?
-        ENERGY_STAR_Score = benchmark['ENERGY_STAR_Score__c']
-        new_ENERGY_STAR_Score = 20
-        args = {"ENERGY_STAR_Score__c": new_ENERGY_STAR_Score}
+        energy_star_score = benchmark["ENERGY_STAR_Score__c"]
+        new_energy_star_score = 20
+        args = {"ENERGY_STAR_Score__c": new_energy_star_score}
         bench_updated = self.sf.update_benchmark(salesforce_benchmark_id, **args)
         print(f"benchmark updated: {bench_updated}")
 
         # retrieve again to see if it was updated
         bench2 = self.sf.get_benchmark_by_custom_id(salesforce_benchmark_id)
-        assert bench2['ENERGY_STAR_Score__c'] == new_ENERGY_STAR_Score
+        assert bench2["ENERGY_STAR_Score__c"] == new_energy_star_score
 
         # restore value
-        args['ENERGY_STAR_Score__c'] = ENERGY_STAR_Score
+        args["ENERGY_STAR_Score__c"] = energy_star_score
         self.sf.update_benchmark(salesforce_benchmark_id, **args)
